refactor: update storage management to use @wxt-dev/storage and adjust key prefixes in components
