+++ packages/next-wxt/entrypoints/sidepanel/useWebAgentServer.ts
@@ -1,5 +1,6 @@
 import { WebMcpClient } from '@opentiny/next-sdk'
-import { storage, StorageKeys } from './utils/storage-manager'
+import { storage } from '@wxt-dev/storage'
+import { StorageKeys } from './utils/storage-keys'
 import { createMcpServer } from './mcpServer'
 import { AGENT_ROOT } from './const'
 
@@ -21,10 +22,10 @@ export const useWebAgentServer = async (): Promise<string> => {
   // 注意：由于存储是异步的，这里先设为 null，后续从存储加载
   // Note: Since storage is async, set to null initially, load from storage later
   let latestSessionId: string | null = null
-  
-  // 从存储加载 sessionId
+
+  // 从存储加载 sessionId（使用 @wxt-dev/storage 统一存储接口）
   try {
-    latestSessionId = await storage.getItemAsync<string>(StorageKeys.MCP_SESSION_ID)
+    latestSessionId = (await storage.getMeta(StorageKeys.MCP_SESSION_ID)) as unknown as string | null
   } catch (error) {
     console.warn('[useWebAgentServer] Failed to load stored sessionId:', error)
   }
@@ -45,11 +46,11 @@ export const useWebAgentServer = async (): Promise<string> => {
     onError
   })
 
-  // 处理连接成功
+  // 处理连接成功（使用 @wxt-dev/storage 统一存储接口）
   const handleConnectSuccess = async (sessionId: string, isRetry: boolean = false) => {
     console.log(`【useWebAgentServer】${isRetry ? '重连' : '连接'}成功，sessionId:`, sessionId)
     try {
-      await storage.setItemAsync(StorageKeys.MCP_SESSION_ID, sessionId)
+      await storage.setMeta(StorageKeys.MCP_SESSION_ID, sessionId as unknown as Record<string, unknown>)
     } catch (error) {
       console.error('[useWebAgentServer] Failed to save sessionId:', error)
     }
